requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.31.1
py4j==0.10.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pyspark==4.0.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.7
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
